Update: resovel comments: - moved stats array to the store - rename stats store
